Fix the connection directions with siblings in the story flow
Previously, the nodes were connected to their previous node
all the time which is incorrect. That should only be done on the
last node in the story. Because the connections must go forward,
the code to do the connection had to moved up to StoryComponent,
since at execution time, the StoryRowComponents would not have access
to the next row since it doesn't exist yet.